requirements.txt
selenium~=4.4.0
webdriver-manager~=3.8.3
PyYAML~=6.0